+++ src/components/toast/global_toast_list.test.js
@@ -7,12 +7,35 @@ import { EuiGlobalToastList } from './global_toast_list';
 describe('EuiGlobalToastList', () => {
   test('is rendered', () => {
     const component = render(
-      <EuiGlobalToastList {...requiredProps}>
-        <div>hi</div>
-      </EuiGlobalToastList>
+      <EuiGlobalToastList
+        {...requiredProps}
+        dismissToast={() => {}}
+        toastLifeTimeMs={1000}
+      />
     );
 
-    expect(component)
-      .toMatchSnapshot();
+    expect(component).toMatchSnapshot();
+  });
+
+  describe('props', () => {
+    describe('toasts', () => {
+      it('is rendered', () => {
+        const toats = [
+          { id: 'A', title: 'A' },
+          { id: 'B', title: 'B' },
+          { id: 'C', title: 'C' },
+        ];
+
+        const component = render(
+          <EuiGlobalToastList
+            toasts={toats}
+            dismissToast={() => {}}
+            toastLifeTimeMs={1000}
+          />
+        );
+
+        expect(component).toMatchSnapshot();
+      });
+    });
   });
 });