+++ src/components/context_menu/context_menu_panel.js
@@ -88,6 +88,8 @@ export class EuiContextMenuPanel extends Component {
     ) {
       if (e.keyCode === cascadingMenuKeyCodes.LEFT) {
         if (this.props.showPreviousPanel) {
+          e.preventDefault();
+          e.stopPropagation();
           this.props.showPreviousPanel();
 
           if (this.props.onUseKeyboardToNavigate) {