+++ src-docs/src/views/toast/toast_list.js
@@ -1,22 +1,16 @@
 import React, {
-  cloneElement,
   Component,
+  Fragment,
 } from 'react';
 
 import {
   EuiGlobalToastList,
-  EuiGlobalToastListItem,
   EuiLink,
-  EuiToast,
 } from '../../../../src/components';
-
-const TOAST_LIFE_TIME_MS = 4000;
-const TOAST_FADE_OUT_MS = 250;
-let toastIdCounter = 0;
-const timeoutIds = [];
 
 let addToastHandler;
 let removeAllToastsHandler;
+let toastId = 0;
 
 export function addToast() {
   addToastHandler();
@@ -39,49 +33,18 @@ export default class extends Component {
   }
 
   addToast = () => {
-    const {
-      toast,
-      toastId,
-    } = this.renderRandomToast();
+    const toast = this.getRandomToast();
 
     this.setState({
       toasts: this.state.toasts.concat(toast),
     });
-
-    this.scheduleToastForDismissal(toastId);
   };
 
-  scheduleToastForDismissal = (toastId, isImmediate = false) => {
-    const lifeTime = isImmediate ? TOAST_FADE_OUT_MS : TOAST_LIFE_TIME_MS;
-
-    timeoutIds.push(setTimeout(() => {
-      this.dismissToast(toastId);
-    }, lifeTime));
-
-    timeoutIds.push(setTimeout(() => {
-      this.startDismissingToast(toastId);
-    }, lifeTime - TOAST_FADE_OUT_MS));
+  removeToast = (removedToast) => {
+    this.setState(prevState => ({
+      toasts: prevState.toasts.filter(toast => toast.id !== removedToast.id),
+    }));
   };
-
-  startDismissingToast(toastId) {
-    this.setState({
-      toasts: this.state.toasts.map(toast => {
-        if (toast.key === toastId) {
-          return cloneElement(toast, {
-            isDismissed: true,
-          });
-        }
-
-        return toast;
-      }),
-    });
-  }
-
-  dismissToast(toastId) {
-    this.setState({
-      toasts: this.state.toasts.filter(toast => toast.key !== toastId),
-    });
-  }
 
   removeAllToasts = () => {
     this.setState({
@@ -89,20 +52,11 @@ export default class extends Component {
     });
   };
 
-  componentWillUnmount() {
-    timeoutIds.forEach(timeoutId => clearTimeout(timeoutId));
-  }
-
-  renderRandomToast = () => {
-    const toastId = (toastIdCounter++).toString();
-    const dismissToast = () => this.scheduleToastForDismissal(toastId, true);
-
-    const toasts = [
-      (
-        <EuiToast
-          title="Check it out, here's a really long title that will wrap within a narrower browser"
-          onClose={dismissToast}
-        >
+  getRandomToast = () => {
+    const toasts = [{
+      title: `Check it out, here's a really long title that will wrap within a narrower browser`,
+      text: (
+        <Fragment>
           <p>
             Here&rsquo;s some stuff that you need to know. We can make this text really long so that,
             when viewed within a browser that&rsquo;s fairly narrow, it will wrap, too.
@@ -110,59 +64,54 @@ export default class extends Component {
           <p>
             And some other stuff on another line, just for kicks. And <EuiLink href="#">here&rsquo;s a link</EuiLink>.
           </p>
-        </EuiToast>
-      ), (
-        <EuiToast
-          title="Download complete!"
-          color="success"
-          onClose={dismissToast}
-        >
-          <p>
-            Thanks for your patience!
-          </p>
-        </EuiToast>
-      ), (
-        <EuiToast
-          title="Logging you out soon, due to inactivity"
-          color="warning"
-          iconType="user"
-          onClose={dismissToast}
-        >
+        </Fragment>
+      ),
+    }, {
+      title: 'Download complete!',
+      color: 'success',
+      text: (
+        <p>
+          Thanks for your patience!
+        </p>
+      ),
+    }, {
+      title: 'Logging you out soon, due to inactivity',
+      color: 'warning',
+      iconType: 'user',
+      text: (
+        <Fragment>
           <p>
             This is a security measure.
           </p>
           <p>
             Please move your mouse to show that you&rsquo;re still using Kibana.
           </p>
-        </EuiToast>
-      ), (
-        <EuiToast
-          title="Oops, there was an error"
-          color="danger"
-          iconType="help"
-          onClose={dismissToast}
-        >
-          <p>
-            Sorry. We&rsquo;ll try not to let it happen it again.
-          </p>
-        </EuiToast>
+        </Fragment>
       ),
-    ];
+    }, {
+      title: 'Oops, there was an error',
+      color: 'danger',
+      iconType: 'help',
+      text: (
+        <p>
+          Sorry. We&rsquo;ll try not to let it happen it again.
+        </p>
+      ),
+    }];
 
-    const toast = (
-      <EuiGlobalToastListItem key={toastId}>
-        {toasts[Math.floor(Math.random() * toasts.length)]}
-      </EuiGlobalToastListItem>
-    );
-
-    return { toast, toastId };
+    return {
+      id: toastId++,
+      ...toasts[Math.floor(Math.random() * toasts.length)],
+    };
   };
 
   render() {
     return (
-      <EuiGlobalToastList>
-        {this.state.toasts}
-      </EuiGlobalToastList>
+      <EuiGlobalToastList
+        toasts={this.state.toasts}
+        dismissToast={this.removeToast}
+        toastLifeTimeMs={6000}
+      />
     );
   }
 }